backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-core
requests
python-dotenv
slowapi
python-dotenv